Moved gtk_clock_start() from new_game() to look_cell() ... I'm not sure
Moved gtk_clock_start() from new_game() to look_cell() ... I'm not
sure that this is the best place to put it, but the effect this has is
to make the clock not start until the player actually clicks the first
cell.  It's hardly fair to start the timer until the player starts
playing!