Gnometris renders with Clutter in an extremely rudamentary way, as a first milestone on the way to a fully-canvas driven game engine. Right now, it's just two ClutterCairo surfaces, background and foreground, which are used as compositors. These two actors will end up staying: background for--eh--the background, and foreground for display of Pango text. An intermediate layer will be introduced on which game pieces are actually animated, rotated and exploded.
Sadly, even this stupid way of rendering is faster than the old method.

This pre-alpha 'leaks' memory like gang-busters, also. I do not know if it is
merely a i915 driver bug or something more fundamental. It's not a 'leak' in the
traditional sense: all the textures that are 'leaked' are still reachable in the
i915 driver code. It leaks on the order of 1MB of memory per frame of animation.
This will be fixed either by doing things the Right Way or mitigating i915 leaks
by avoiding the use of texture uploads at all costs (a tile cache).

svn path=/trunk/; revision=8397