Redirect translations from gnometris/renderer.cpp to noclutter version.
Clutter work has moved the older rendering code in to
renderer-noclutter.cpp. So we should have the translators working on
that file and not the renderer.cpp file that will be disabled-by-default
for 2.26. This also allows string changes in renderer.cpp to continue on
trunk during the freeze.

Changes to PO files were done using a sed.

svn path=/trunk/; revision=8627