Factor out all calls to redraw so that we do not use cairo for animation. This change eliminates more surface oriented drawing logic but preserves backward compatibility with the old engine.
This also removes the 'target' drawing code which will have to be done
entirely differently in Clutter. For now, the UI elements remain in
order to avoid and warning about string changes during the string
freeze. It will be reimplemented eventually anyway.

Finally, this commit disables building -noclutter binary variant when
in the HAVE_CLUTTER case.

svn path=/trunk/; revision=8647