Make renderer render cells and cache instead of the whole field.
This set of changes does away with the concept of the renderer as a
slave that draws the entire field by itteration through field block
data. Instead, the renderer has been paired down to knowledge of only
how to draw an individual cell in anticipation of renderer handling its
own cache of the seven block colors and providing textures to blockops
on demand when requested.

Also, this commit splits blockops out in to to pieces so that Actor
movement logic can be kept in blockops.

svn path=/trunk/; revision=8648