Implement a Clutter tile cache in Renderer
This commit implements a tile cache using CLUTTER_CAIRO as the source
surface. The game field uses a CLUTTER_CLONE_TEXTURE to set the cairo
surface as its parent. Resize or theme change events automatically
destroy the cache and recreate it using the new level of precision.
Additionally, theme changes completely destroy the renderer object and
recreate it.

I intentionally did not reuse the code in use for Aisleriot as I have no
idea what its doing and this seems incredibly simple by comparison

Also, this commit fixes a few whitespace things and moves NCOLOURS to
blocks.h so that it is more globally accessible.

svn path=/trunk/; revision=8649