Bump the min. Clutter version to 0.8.7 and require Clutter-Gtk 0.8.2
Clutter 0.8.6 has a critical Behaviour bug where Behaviours with no
associated actors or knots will cause a segmentation fault. We use this
type of Behaviour to avoid constantly alloc and decallocing Behaviours
for use for moving the pieces around on the board and also to avoid a
flood of callbacks when movement is completed. (They all share a common
timeline.)

Our in-tree Clutter-Gtk also appears to be incompatible, so it is
deprecated with at least Gnometris. Might need to do this for Aisleriot,
as well.

svn path=/trunk/; revision=8699