Massive rewrite of movement animation and performance improvements
This change reduces the number of callbacks we end up handling by a very
significant amount. The strategy for movement of pieces that are in the
FALLING state is to use a static Behaviours array that can be stopped,
adjusted by appending new knots. And the resumed. This has the huge
benefit of making it possible accelerate the piece to a very rapid pace
making sure that it never takes more than 60ms to reach the final
resting place.

svn path=/trunk/; revision=8701