[aisleriot] Avoid drawing unexposed cards that are being moved
When detecting animations it would previously only check exposed cards
at the top of the slot. However sometimes unexposed cards can be moved
as well, such as when the waste slot is flipped over to the draw slot
in Canfield. The unexposed cards were just being moved instantly while
the exposed cards were animated. Now it keeps track of the number of
exposed cards that have moved and doesn't draw them during an
animation so that it should appear as if they are being moved with the
exposed cards.

svn path=/trunk/; revision=8813