[aisleriot-slot-renderer] Allow the animation to specify the old card
Instead of specifying whether the old card was flipped, the whole
value of the old card is now passed. It will animate turning over the
cards if they are different.

AisleriotBoard uses this to set the bottom card to be the bottom most
unexposed card. This gives a better appearance when flipping over a
whole slot.

svn path=/trunk/; revision=8815