Beginning of port to clutter-0.9 API and refactoring of animation code
This doesn't actually work yet but at least it passes build for the
purposes of our 2.27.1 release. The next commit will have something in
it which resembles the 2.26.x releases.

The refactoring is an attempt to encapsulate all animation to
block-level tracking rather than having it spread between the field and
the block objects.