Add annoying "bastard" mode
Sometimes, gnometris is not evil enough.  This adds inpopular "Bastard" [1]
mode to gnometris.  When enabled, it intentionally provides the user
unsuitable pieces making the play a very frustrating experience.  It also
annoys the user by displaying the most suitable piece instead of the next
to come.

In order to make it a bit fun, it is not that evil.  It more-or-less
picks from three worst pieces giving the highest chance to the worst
one so that the game is actually playable.

The idea comes from "bastet" [2], which stands for "Bastard Tetris".
This shares the algorithm, but no code.

[1] http://games.slashdot.org/games/05/04/24/1259241.shtml?tid=208&tid=106
[2] http://fph.altervista.org/prog/bastet.shtml

Signed-off-by: Jason D. Clinton <me@jasonclinton.com>