First pass at updating Gnome Games use of Clutter 0.9.x API to master
Clutter 0.9.3 and 0.9.4 are markedly incompatible though--as 0.9.4 is a
release candidate--there should be no additional breakages in API. This
first pass gets things compiling but the following is the case:

A) aisleriot displays the background but no cards; likely a Units issue
B) gnometris, though no major changes were made, displays no pieces

A follow-up commit will resolve these issues and bump the min.
requirement to 0.9.4 (unreleased).